+++ compiler/flags/GNU_Fortran.cmake
@@ -1,5 +1,5 @@
-if (CMAKE_Fortran_COMPILER_VERSION VERSION_LESS 8.3)
-  message(FATAL_ERROR "${CMAKE_Fortran_COMPILER_ID} version must be at least 8.3!")
+if (CMAKE_Fortran_COMPILER_VERSION VERSION_LESS 10.3)
+  message(FATAL_ERROR "${CMAKE_Fortran_COMPILER_ID} version must be at least 10.3!")
 endif()
 
 set (FOPT0 "-O0")
@@ -131,6 +131,15 @@ elseif (${proc_description} MATCHES "EPYC")
   set (GNU_TARGET_ARCH "znver2")
   set (GNU_NATIVE_ARCH "native")
   set (NO_FMA "-mno-fma")
+elseif (${proc_description} MATCHES "Hygon")
+  # Tests on a Hygon showed it returned 'nocona' for the native arch
+  # see https://github.com/geoschem/GCHP/issues/391
+  # Intel was happy with AVX2, but this test was done with GCC 10
+  # Perhaps later versions of GCC would return a different value?
+  # Until then, nocona is definitely safe. Pentium 4!
+  set (GNU_TARGET_ARCH "nocona")
+  set (GNU_NATIVE_ARCH "native")
+  set (NO_FMA "-mno-fma")
 elseif (${proc_description} MATCHES "Intel")
   set (GNU_TARGET_ARCH "haswell")
   set (GNU_NATIVE_ARCH "native")
@@ -142,6 +151,24 @@ elseif ( ${CMAKE_HOST_SYSTEM_PROCESSOR} STREQUAL "x86_64" )
   set (GNU_NATIVE_ARCH "native")
 else ()
   message(FATAL_ERROR "Unknown processor. Please file an issue at https://github.com/GEOS-ESM/ESMA_cmake")
+endif ()
+
+if (APPLE)
+  # Determine whether we need to add link options for version 15+ of the Apple command line utilities
+  execute_process(COMMAND "pkgutil"
+                          "--pkg-info=com.apple.pkg.CLTools_Executables"
+                  OUTPUT_VARIABLE TEST)
+  string(REGEX REPLACE ".*version: ([0-9]+).*" "\\1" CMDLINE_UTILS_VERSION ${TEST})
+  message(STATUS "Apple command line utils major version is '${CMDLINE_UTILS_VERSION}'")
+  if (${CMDLINE_UTILS_VERSION} VERSION_GREATER 14)
+    message(STATUS "Adding link options '-Wl,-ld_classic'")
+    add_link_options(-Wl,-ld_classic)
+  endif ()
+  #
+  # Also, if our C compiler is Apple Clang, we need to pass -Wno-implicit-int to our C flags
+  if (CMAKE_C_COMPILER_ID MATCHES "Clang")
+    set (CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -Wno-implicit-int")
+  endif ()
 endif ()
 
 ####################################################