+++ esma_support/esma_support.cmake
@@ -8,6 +8,7 @@ include (esma_add_library)
 include (esma_generate_automatic_code)
 include (esma_create_stub_component)
 include (esma_fortran_generator_list)
+include (esma_add_fortran_submodules)
 
 # Testing
 include (esma_enable_tests)